extrac common systemd config
